requirements_dev.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
billiard==3.6.4.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.7
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
celery==5.2.7
redis==4.5.1
gunicorn==20.1.0
sqlalchemy
requests
beautifulsoup4